refactor: 書き込みを保存するテーブルに共通する部分を1つのモデルにまとめる
