+++ app/Models/ThreadModel.php
@@ -1,0 +1,97 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\SoftDeletes;
+
+class ThreadModel extends Model
+{
+    use HasFactory,
+        SerializeDate,
+        SoftDeletes;
+
+    /**
+     * 接続するデータベース
+     *
+     * @link https://readouble.com/laravel/9.x/ja/eloquent.html
+     * @see config/database.php
+     *
+     * @var string
+     */
+    protected $connection = 'mysql';
+
+    /**
+     * マスアサインメント可能な属性
+     *
+     * ここに登録している属性にはデータの挿入・更新が出来る．
+     *
+     * @link https://readouble.com/laravel/9.x/ja/eloquent.html
+     *
+     * @var string[]
+     */
+    protected $fillable = [
+        'hub_id',
+        'user_id',
+        'message_id',
+        'message',
+    ];
+
+    /**
+     * キャストすべき属性
+     *
+     * @link https://readouble.com/laravel/9.x/ja/eloquent-serialization.html
+     *
+     * @var array
+     */
+    protected $casts = [
+        'created_at' => 'datetime:Y-m-d H:i:s',
+        'updated_at' => 'datetime:Y-m-d H:i:s',
+        'deleted_at' => 'datetime:Y-m-d H:i:s',
+    ];
+
+    /**
+     * lecture thread に関連する thread image path を取得します．
+     * 1：1
+     *
+     * @link https://readouble.com/laravel/9.x/ja/eloquent-relationships.html
+     */
+    public function thread_image_path()
+    {
+        return $this->hasOne(ThreadImagePath::class);
+    }
+
+    /**
+     * lecture thread に関連する like を取得する．
+     * 1：多
+     *
+     * @link https://readouble.com/laravel/9.x/ja/eloquent-relationships.html
+     */
+    public function likes()
+    {
+        return $this->hasMany(Like::class);
+    }
+
+    /**
+     * lecture thread を所有する hub を取得します．
+     * 多：1
+     *
+     * @link https://readouble.com/laravel/9.x/ja/eloquent-relationships.html
+     */
+    public function hub()
+    {
+        return $this->belongsTo(Hub::class);
+    }
+
+    /**
+     * lecture thread を所有する user を取得します．
+     * 多：1
+     *
+     * @link https://readouble.com/laravel/9.x/ja/eloquent-relationships.html
+     */
+    public function user()
+    {
+        return $this->belongsTo(User::class);
+    }
+}
